pyproject.toml
[tool.black]
line-length = 88
exclude = '''
/(
  | migrations
  | .venv
)/
'''

[tool.flake8]
ignore = ["F405", "E501", "W503", "E731", "E741",]
max-line-length = 88
exclude = [
    ".git", "migrations", "node_modules", "docs", ".venv", "frontend"
]
statistics = false

[tool.isort]
line_length = 88
skip_glob = "*/migrations/.py"
multi_line_output = 3
profile = "black"
filter_files = true
include_trailing_comma = true
force_grid_wrap = 2

[tool.poetry]
name = "Django core project"
version = "0.1.0"
description = ""
authors = ["Piwero <piwero@gmail.com>"]

[tool.poetry.dependencies]
Django = "^3.2.5"
django-allauth = "^0.45.0"
python = "^3.9"
python-decouple = "^3.4"
django-crispy-forms = "^1.12.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
ipdb = "^0.13.9"
isort = "^5.9.3"
flake9 = "3.8.3"
pre-commit = "^2.13.0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
